requirements.txt
attrs==17.3.0
backcall==0.1.0
decorator==4.1.2
Django==2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.3
ipython-genutils==0.2.0
jedi==0.11.0
Markdown==2.6.9
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
